+++ src/mec.node.js
@@ -135,6 +135,13 @@ mec.node = {
 
             return obj;
         },
+        // analysis methods
+        force() { return {x:this.Qx,y:this.Qy}; },
+        vel() { return {x:this.xt,y:this.yt}; },
+        acc() { return {x:this.xtt,y:this.ytt}; },
+        forceAbs() { return Math.hypot(this.Qx,this.Qy); },
+        velAbs() { return Math.hypot(this.xt,this.yt); },
+        accAbs() { return Math.hypot(this.xtt,this.ytt); },
         // interaction
         get isSolid() { return true },
         get sh() { return this.state & g2.OVER ? [0, 0, 10, mec.hoveredElmColor] : this.state & g2.EDIT ? [0, 0, 10, mec.selectedElmColor] : false; },